requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
Pillow
pymongo
python-dotenv